+++ Assets/_Scripts/Player/PlayerSkills/Behaviors/Movement/ISkillMovement.cs
@@ -7,5 +7,5 @@ using UnityEngine;
 /// </summary>
 public interface ISkillMovement
 {
-    void MoveSkill(GameObject skillObject, Vector3 targetPos, int speed);
+    void MoveSkill(GameObject skillObject, Vector3 targetPos, float speed);
 }
